pyproject.toml
[tool.poetry]
name = "marketing-crew"
version = "0.1.0"
description = ""
authors = ["ori"]

[tool.poetry.dependencies]
python = ">=3.10.0,<4.0"
crewai = "^0.11.0"
setuptools = "^69.1.0"
duckduckgo-search = "^4.5.0"
python-dotenv = "^1.0.1"
langchain-community = "^0.0.22"


[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"